requirements.txt
asgiref==3.3.1
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1
djangorestframework
djangorestframework-simplejwt
psycopg2
knox
drf_yasg
pytest-django
gunicorn
coverage
random_username
djangorestframework-simplejwt